server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# bcrypt 4+ breaks passlib 1.7.4 (wrap bug / __about__ checks)
bcrypt<4.0.0,>=3.2.0
httpx[http2]>=0.27.0
h2>=4.0.0
python-multipart>=0.0.12
aioslsk>=1.6.3
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
slowapi>=0.1.9
